+++ mine/src/main.rs
@@ -86,7 +86,7 @@ fn main() {
         .push(pepper_plugin_lsp::DEFAULT_BINDINGS_CONFIG);
     config.static_configs.push(ResourceFile {
         name: "my.pepper",
-        content: "map-normal ga {{: goto-alternate-buffer<enter>}}",
+        content: "map normal ga :<space>goto-alternate-buffer<enter>",
     });
 
     pepper::run(config);